Revert "Updated Drivers, fixed an error in the Shooter Motor Subsystem"
This reverts commit c571ebbd4cf701d4d8ed73fcc38f5adadcb9e9d2.

+++ src/main/java/frc/robot/subsystems/shooterMotorSubsystem.java
@@ -107,7 +107,7 @@ public void runShooterEncoder() {
 
 }
 
-public double getVelocityError() {
+public int getVelocityError() {
   
   
   System.out.print(" Velocity1 ");
